New file for PublicKey definition and constructor
